feat: add Specialty controllers and update entity with description and image properties
- Added controllers for managing Specialty entity (CRUD operations).
- Updated Specialty entity to include new properties: 'description' and 'image'.
- Enhanced Studio deletion controller to also remove associated images from S3 bucket.

+++ src/main/java/com/luxevision/backend/controller/StudioController.java
@@ -42,7 +42,14 @@ public class StudioController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteStudio(@PathVariable Integer id) {
+
+        Optional<Studio> studioFromDB = studioService.findStudioById(id);
+        if (!studioFromDB.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+
         try {
+            s3Service.deleteObject(null, studioService.findStudioById(id).get());
             studioService.deleteStudioById(id);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         } catch (Exception e) {
